chore: misc clean up
